Update patch set 1
Patch Set 1: Do not submit

I'd rather this be fixed in the app, there is no need for the extra setting.

The app's correct behavior should be to not set the value if it's already set.

If this affects how some devices work, I'd rather check it on the kernel level than adding another option.

Patch-set: 1